requirements.txt
flask
streamlit
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
pypdf
python-dotenv
tiktoken